Добавлено получение флаг поинтов для сервиса

+++ src/main/java/ru/hits/attackdefenceplatform/public_interface/team/TeamInfoDto.java
@@ -12,7 +12,7 @@ public record TeamInfoDto(
         Long userCount,
         Long membersCount,
         Integer place,
-        Integer points,
+        Double points,
         Boolean canJoin,
         Boolean isMyTeam,
         Boolean canLeave,
